+++ src/documentFormattingEditProvider.ts
@@ -1,16 +1,18 @@
 import * as vscode from 'vscode';
 
-type TextDocumentFormatter = (doc: vscode.TextDocument) => Promise<string>;
-export const createFormattingProvider: (createTextDocumentFormatter: () => TextDocumentFormatter) => vscode.DocumentFormattingEditProvider =
-    (createTextDocumentFormatter) => {
+type TextDocumentFormatter = (doc: vscode.TextDocument) => Promise<string | undefined>;
+
+export const createFormattingProvider: (textDocumentFormatter: TextDocumentFormatter) => vscode.DocumentFormattingEditProvider =
+    (textDocumentFormatter) => {
         return {
             provideDocumentFormattingEdits: async (document, _formattingOptions, _token) => {
                 try {
-                    const f = createTextDocumentFormatter();
-                    const newContent = await f(document);
-                    return textToFullDocumentReplacement(document)(newContent);
+                    const newContent = await textDocumentFormatter(document);
+                    if (!newContent) {
+                        return;
+                    }
+                    return textToFullDocumentReplacement(document, newContent);
                 } catch (e) {
-                    console.error("Formatter failed", e);
                     vscode.window.showErrorMessage(`${e}`);
                     return null;
                 }
@@ -18,11 +20,11 @@ export const createFormattingProvider: (createTextDocumentFormatter: () => TextD
         };
     };
 
-const textToFullDocumentReplacement: (document: vscode.TextDocument) => (text: string) => vscode.TextEdit[] =
-    (document) =>
-        (text) => {
-            const firstLine = document.lineAt(0);
-            const lastLine = document.lineAt(document.lineCount - 1);
-            const textRange = new vscode.Range(firstLine.range.start, lastLine.range.end);
-            return [vscode.TextEdit.replace(textRange, text)];
-        };
+
+const textToFullDocumentReplacement: (document: vscode.TextDocument, text: string) => vscode.TextEdit[] =
+    (document, text) => {
+        const firstLine = document.lineAt(0);
+        const lastLine = document.lineAt(document.lineCount - 1);
+        const textRange = new vscode.Range(firstLine.range.start, lastLine.range.end);
+        return [vscode.TextEdit.replace(textRange, text)];
+    };