+++ src/externalTextDocumentFormattingProvider.ts
@@ -1,28 +1,24 @@
 import * as vscode from 'vscode';
 import { createFormattingProvider } from './documentFormattingEditProvider';
-import * as tdf from './textDocumentFormatter';
 import * as etf from './externalTextFormatter';
 import * as path from 'path';
 import * as process from 'process';
+import * as config from './configuration';
 
-export const create: () => vscode.DocumentFormattingEditProvider = () =>
-    createFormattingProvider(() =>
-        tdf.createTextDocumentFormatter((td) =>
-            etf.externalTextFormatter(getSettingsForDocument(td))
-        )
+export const create: () => vscode.DocumentFormattingEditProvider =
+    () => createFormattingProvider(
+        async (td) => {
+            const formatter = etf.externalTextFormatter(getSettingsForDocument(td));
+            return await formatter(td.getText());
+        }
     );
 
-export interface ExternalFormatterSettings {
-    command: string;
-    arguments: string[];
-}
 
 const getSettingsForDocument: (doc: vscode.TextDocument) => etf.FormatterSettings =
     (doc) => {
-        const config = vscode.workspace.getConfiguration('externalFormatters');
-        const forLanguage = config.get<ExternalFormatterSettings>(doc.languageId);
+        const forLanguage = config.getLanguageConfiguration(doc.languageId);
         if (!forLanguage) {
-            throw new Error(`Invalid settings for "externalFormatters.${doc.languageId}"`);
+            throw new Error(`Bad settings for externalFormatters.${doc.languageId}`);
         }
         return {
             command: forLanguage.command,
@@ -31,9 +27,16 @@ const getSettingsForDocument: (doc: vscode.TextDocument) => etf.FormatterSetting
         };
     };
 
-const rootFolder: () => string | undefined = () => vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0].uri.fsPath;
 
-const documentDir: (td: vscode.TextDocument) => string | undefined = (textDocument: vscode.TextDocument) => {
-    if (textDocument.uri.scheme !== 'file') { return; }
-    return path.dirname(textDocument.uri.fsPath);
-};
+const rootFolder: () => string | undefined =
+    () => (
+        vscode.workspace.workspaceFolders &&
+        vscode.workspace.workspaceFolders[0].uri.fsPath
+    );
+
+
+const documentDir: (td: vscode.TextDocument) => string | undefined =
+    (textDocument: vscode.TextDocument) => {
+        if (textDocument.uri.scheme !== 'file') { return; }
+        return path.dirname(textDocument.uri.fsPath);
+    };