+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
-All notable changes to the "call-external-formatter-on-save" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+
+- Dedicaded output channel for external formatter stdout en stderr
+- Improved logging
+- Re-register language formatter upon config changes
+
+## [0.0.2] - 2019-12-23
 
 - Initial release